Create Stephen Jennifer Ebubechukwu- 2024714070
This Python program implements a grading system using IF and elif statements. It takes a student's score as an input and prints the corresponding grade based on the following grade:
- 70- 69: Grade A
- 60-69: Grade B
- 50-59: Grade C
- 45-49: Grade D
- 40-44: Grade E
- 20-39: Grade F

For each valid score, the program also prints "YES" to confirm the grading was successful. If the score is outside the valid range, it prints invalid score and "NO".

Created by; Stephen Jennifer Ebubechukwu 

Reg no: 2024714070